Update patch set 3
Patch Set 3:

(4 comments)

Hi!
During a previous project we discussed to not have a dedicated storage section but instead to define storage-related requirements for each layer of the architecture. In this context, I proposed these storage requirements in this virtualization-layer section because I was referring to "software-based" storage aspects which are located in the host OS or hypervisor, such as distributed storage services (e.g. CEPH), instead of in a dedicated hardware storage array.

Do you agree to collect software-related storage requirements in this section?

Patch-set: 3
Label: Verified=0