Update patch set 3
Patch Set 3:

Line 129: By "specify for each class of virtual resources the associated HA handling...mechanisms...", I'm not sure if this implies that the deployment of an VNF specifies the specific mechanism, or if the operator designs a specific mechanism per (one or many) VNFs. Surely the former is too limiting since it creates implementation-specific dependencies on the VNF deployment, whereas specifying abstracted service requirements would be fine. OTOH, if the latter meaning is intended, then I don't know why this would be a requirement in a specification as opposed to a required behavior of the operator.

Line 184: Table 2 needs to be reconciled with the table in Section 1.4. These SAL requirements are simply not achievable in a realistic fashion.

Line 286: What if the NFV is recovered in another NFVI? Then the NFVI must not maintain the number of VNFs. Perhaps what is missing is a requirement that says whether applies to local recoverability, remote recoverability, or both, depending on how the NF is configured.

Line 288: Where should the specification of VNF/VNFM vs NFVI recovery handling be specified? If it's in the design of the service by the operator, then sure, but that's not a requirement, is it? If it's in the deployment of a VNF into an NFVI, then what if the requested method is not available? 

Line 288 & 292 & 310: Recovery capability mechanisms are various HA implementations (I noticed the hypervisor was not listed as a mechanism). I'm not sure what the intent is of these requirements. Is the intent to tell vendors what the implementation mechanisms are? Many of those (of course) are proprietary. If the intent is to provide guidance to the upstream community that OPNFV needs an open source implementation of said mechanism, then does that mean contributor is offering to submit that open source mechanism?

Line 322: New requirement - VIM availability: The VIM shall be recoverable in the local environment within 2 minutes of failure detection.

Line 322: New requirement - VIM availability: The VIM shall be recoverable in a geographically separate environment within 2 minutes of site failure detection.

Line 322: New requirement - VIM rollback: The VIM shall be able to roll back the most recent changes made to the NFVI (patches, updates, etc).

Line 322: New requirement - VIM backup: the VIM shall have its configuration backed up to a remote location after every series of configuration changes. (A "series" being operator-determined, or after 1 hour of the last change.)

My apologies if these comments don't fit the desired form. This is as close as I could figure out how to inject these.

Patch-set: 3
Reviewer: Gerrit User 830 <830@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Verified=0