Update patch set 4
Patch Set 4: Code-Review-1

(7 comments)

Hi Qiao,
I think some discussions would be good in this section as well and group the requirements as applicable to the discussion. For example, it's not clear if the requirements related to storage would apply to all hosts of the NFVI.
Also requirements at the NFVI level would need to be decomposed how they apply to the entities discussed, e.g. host OS, hypervisor.
Thanks,
Maria

Patch-set: 4
Label: Code-Review=-1