Update patch set 3
Patch Set 3:

(6 comments)

Hi, Maria. Thank you for the update. I read it through and I think most of it is quite good. I still have a few questions which I proposed in the rst.
I will call the whole team to review this and see if we have more comments. If not, I think we are almost there to have this doc merged.
In the meantime, I suggest we build another rst for the other usecases, so that it might be easier for us to review and update this scenario doc. We can talk about this after we discuss with the multisite project this week.
thank you!

Patch-set: 3