Update patch set 1
Patch Set 1:

(5 comments)

Hi, Edgar. thank you for this great work. Since I am not quite familiar with the storage system, I found the first session difficult to understand. I think a topology for the storage system deployment might be helpful for me to understand which exactly module you are refering to in the context.

Patch-set: 1
Label: Code-Review=0