Update patch set 5
Patch Set 5: Code-Review+1

This is helpful for implementors of VNFs. But the conclusion has to be followed. Bevore using this functions on a OPNFV level, the concept has to be synchronized with other projects.
I would like to support Fu Qiao's proposal for a step by step release. start with functions that havclearly been designed to be used locally

Patch-set: 5
Reviewer: Gerrit User 1334 <1334@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=+1