Update patch set 1
Patch Set 1:

We add the proposal word according to the discussion within the team in the last meeting before the D release. It was decided that we will put this as a proposal for the D release, and will remove it once we finish discussion with the doctor team and the MANO WG, making sure the scheme mentioned in this doc coordinate with the related schemes in OPNFV. Therefore, I suggest to not remove the poposal word, but we will update this doc for future releases.

Patch-set: 1
Reviewer: Gerrit User 314 <314@bbac25d2-bf60-4904-9ba8-a72fc000d6c5>
Label: Code-Review=0