Update patch set 1
Patch Set 1:

Ray suggests to remove the poposal word, but to add the word of "work on progress", which makes it more suitable for a release doc. I think this makes sense to me as well. I will wait and see the other committer's comments. Taking consideration of the deadline of Danube release, I hope people give their review before May 1st. I will make the decision after that and give it +2 or not.

Patch-set: 1