Remove inactive committers
The TSC decided to automatically remove the committers that have been inactive
in gerit for a year or more. If any of the people removed by this patch are
contributing to the project in other ways, and the PTL would like to keep the
gerrit rights for them, please ammend this patch.

JIRA: TSC-14

Change-Id: I537ef9a3a6c7c836c92e263cc4f2b8957130e8c8
Signed-off-by: Cristina Pauna <cristina.pauna@enea.com>